% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{AmpGram_predictions}
\alias{AmpGram_predictions}
\title{Prediction of antimicrobial peptides}
\format{
A list of length three: random forest for 10-mer analysis, 
random forest for predictions of AMPs, and a vector of important n-grams.
}
\description{
Predictions made with the AmpGram methods.
}
\keyword{datasets}
