% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_AmpGramModel}
\alias{install_AmpGramModel}
\title{Install AmpGramModel package containing model for AMP prediction}
\usage{
install_AmpGramModel()
}
\description{
Installs AmpGramModel package containing model required for prediction
of antimicrobial peptides. Due to large size of our model and file size 
limit on CRAN, it needs to be stored in the external repository.
See readme for more information or in case of installation problems.
}
