% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.ampgram_model}
\alias{predict.ampgram_model}
\title{Predict amyloids}
\usage{
\method{predict}{ampgram_model}(object, newdata, ...)
}
\arguments{
\item{object}{\code{ampgram_model} object.}

\item{newdata}{\code{list} of sequences (for example as given by
\code{\link[biogram]{read_fasta}}).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Recognizes amyloids using AmpGram algorithm.
}
