% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmpGram_gui.R
\name{AmpGram_gui}
\alias{AmpGram_gui}
\title{AmpGram Graphical User Interface}
\usage{
AmpGram_gui()
}
\description{
Launches graphical user interface that predicts presence of 
antimicrobial peptides.
}
\section{Warning }{
 Any ad-blocking software may cause malfunctions.
}

