% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_AMPs.R
\name{get_AMPs}
\alias{get_AMPs}
\title{Get putative antimicrobial peptides}
\usage{
get_AMPs(x)
}
\arguments{
\item{x}{AmpGram predictions for a single protein}
}
\value{
a data.frame with sequences recognized as antimicrobial peptides.
}
\description{
Function gets sequences recognized as antimicrobial peptides and returns as data.frame.
}
\examples{
data(AmpGram_predictions)
get_AMPs(AmpGram_predictions[[2]])
}
