% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.ampgram_model}
\alias{predict.ampgram_model}
\title{Predict antimicrobial peptides}
\usage{
\method{predict}{ampgram_model}(object, newdata, ...)
}
\arguments{
\item{object}{\code{ampgram_model} object.}

\item{newdata}{\code{list} of sequences (for example as given by
\code{\link[biogram]{read_fasta}} or \code{\link{read_txt}}).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Recognizes antimicrobial peptides using the AmpGram algorithm.
}
\details{
AmpGram requires the external package, AmpGramModel, which 
contains models necessary to perform the prediction. The model 
can be installed using \code{\link{install_AmpGramModel}}.
}
