% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ngrams.R
\name{predict.ampgram_model}
\alias{predict.ampgram_model}
\title{Predict amyloids}
\usage{
\method{predict}{ampgram_model}(object, newdata, ...)
}
\arguments{
\item{object}{\code{ampgram_model} object.}

\item{newdata}{\code{list} of sequences (for example as given by
\code{\link[biogram]{read_fasta}}).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Recognizes amyloids using AmpGram algorithm.
}
\examples{
data(AmpGram_model)
sample_seq <- list(seq1 = c("F", "E", "N", "C", "N", "I", "T", "M", "G", "N", "M", "V", 
                            "R", "H", "I", "R", "W", "Y", "R", "D", "R", "Q", "K", "G", "D", 
                            "Y", "W", "W", "Y", "T", "I", "K", "Y", "S", "M", "A", "M", "I", 
                            "A", "C", "N", "I", "N", "V", "T", "I", "N", "Q", "C", "V"),
                   seq2 = c("Q", "Y", "T", "S", "I", "M", "F", "L", "T", "A", "G", "H", 
                            "L", "A", "P", "W", "D", "R", "W", "C", "R", "S", "L", "T", "T", 
                            "W", "F", "G", "A", "P", "S", "A", "T", "Y", "P", "F", "F", "W", 
                            "E", "P", "E", "D", "I", "I", "I", "K", "P", "N", "T", "A"))
predict(AmpGram_model, sample_seq)
}
